Don't autolink a bare URL if it's part of a previous inline
Previously it was possible to get the renderer to truncate the
output at a point that would cause it out output unbalanced tags,
for ex.

`/r/irc://foo.bar` -> `<a href="/r/irc">/r/irc<<a href="irc://foo[…]`

Not great. This commit attempts to mitigate that by tracking the end
of the last inline and making sure rewinds don't run past it. This is
similar to the approach taken by @vmg in rinku:

https://github.com/vmg/rinku/commit/dc685c15e30d9a02a0144170798538ea4754f0a6