Add an enter_* callbacks for widget toolkit consumers
These callback is useful for widget toolkit users of the render, which
need to insert a widget at when entering (as well as exiting) some of
the block or table types.  This is what we are using in
something-for-reddit which uses the Gtk+ widget toolkit.  For example,
we add a table in the "enter_table" callback, and mark the table as
completed during the "table" callback.